Initial port of 1/2 of twp.instancer to new thing
All this does is cache_tiddlers based on the `store_contents`
dict in an instance module in a named package.

A twibuilder script is included, which means that you don't
have to have a customer cacher script for every package, you
just say `twibuilder <packagename>` in your Makefile or wherever.